Fix int64 cast

+++ internal/fetcher/fetch.go
@@ -27,7 +27,7 @@ func saveIfNewestRelease(artist string, release *itunes.LastRelease) bool {
 		StoreID:    release.ID,
 	})
 	notify.Service.Send(map[string]interface{}{
-		"chatID": 35152258,
+		"chatID": int64(35152258),
 		"message": fmt.Sprintf("new release from '%s' '%s/%s/%d': ",
 			artist, config.Config.Store.URL, config.Config.Store.Region, release.ID),
 	})
